requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
opencv-python==4.8.1.78
pytesseract==0.3.10
# sqlite3 já incluído no Python
pydantic==2.5.0
requests==2.31.0
numpy==1.26.2
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiofiles==23.2.1
psycopg2-binary==2.9.9
jinja2==3.1.2
openai==1.3.0